blog_6/requirements.txt
asgiref==3.5.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.11
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-import-export==2.7.1
django-js-asset==1.2.2
djangorestframework==3.13.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.0
psycopg2==2.9.3
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.0
tzdata==2021.5
xlrd==2.0.1
xlwt==1.3.0
